requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
attrs==25.4.0
azure-core==1.36.0
azure-functions==1.24.0
azure-functions-durable==1.4.0
azure-storage-blob==12.27.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
frozenlist==1.8.0
furl==2.1.4
idna==3.11
importlib_metadata==8.7.0
isodate==0.7.2
MarkupSafe==3.0.3
multidict==6.7.0
numpy==2.2.6
opencv-python==4.12.0.88
opencv-python-headless==4.11.0.86
opentelemetry-api==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
orderedmultidict==1.0.1
propcache==0.4.1
pycparser==2.23
python-dateutil==2.9.0.post0
requests==2.32.5
six==1.17.0
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
yarl==1.22.0
zipp==3.23.0
